Merge branch '494-meshiterator-cast-cell-type-into-celltype-enum-instead-of-using-std-uint8_t' into 'dev'

When iterating through mesh cells, the iterator return the type as a CellType enum instead of a std::uint8_t. It allow to avoid a static_cast to compare the type to the enum.

Closes #494

See merge request sight/sight!417

+++ libs/core/fwDataIO/src/fwDataIO/writer/MeshWriter.cpp
@@ -59,8 +59,7 @@ void MeshWriter::write()
 
     ::fwData::Mesh::csptr mesh = this->getConcreteObject();
     FW_RAISE_IF("Can't convert this Mesh to Trian file",
-                !::fwDataTools::Mesh::hasUniqueCellType(mesh,
-                                                        static_cast<std::uint8_t>(::fwData::Mesh::CellType::TRIANGLE)));
+                !::fwDataTools::Mesh::hasUniqueCellType(mesh, ::fwData::Mesh::CellType::TRIANGLE));
 
     std::fstream file;
     file.open(getFile().string().c_str(), std::fstream::out);
